Added comments explaining what pre-training and fine-tuning does
In particular, clarified why the SdA error (e.g., on a test set) before
and after pre-training is the same and is equal to the pure chance
error. This is because pre-training only trains the underlying sigmoid
layers of the SdA, while the logistic layer on top remains untrained.
The logistic layer on top of SdA only gets trained (in a supervised
mode, of course) in the fine-tuning stage.